Setup of test paseto builders is now done via reflection.
Meta projects are now only required to implement TestBuilders and annotate the class with @Provided.
The class must have an empty constructor. The base testing library will search for classes on the
class path which implement TestBuilders and are annotated with @Provided then call the empty constructor.
This instance is then cached in TestBuilders to avoid the lookup each time.

This is only used for testing.